+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 Anticipated release: September 21st, 2020
 
 #### 🚀 New features
-
 - Updates markup to use semantic `<nav>` and `<main>` where appropriate ([#2502])
+- Change order of fields and field/title names on Cost Allocation and Other Funding + Budget and FFP pages. ([#2322])
+- Other funding amount should be represented as a subtraction on Activity Breakdown table ([#2429])
 
 #### 🐛 Bugs fixed
 
@@ -17,3 +18,5 @@ Anticipated release: September 21st, 2020
 See our [release history](https://github.com/CMSgov/eAPD/releases)
 
 [#2502]: https://github.com/CMSgov/eAPD/issues/2502
+[#2322]: https://github.com/CMSgov/eAPD/issues/2322
+[#2429]: https://github.com/CMSgov/eAPD/issues/2429